// src/types.ts
export type UserRole = 'Super Admin' | 'Admin' | 'User';

export interface UserProfile {
  uid: string;
  email: string;
  displayName: string;
  role: UserRole;
  designation?: string;
  department?: string;
  active: boolean;
  createdAt: string;
}
// ...
export interface SystemSettings {
  organizationName: string;
  organizationLogo: string; // base64 or url
  emailNotifications: boolean;
  reminderDaysBefore: number; // e.g. 3 days
}